refactor: useEditor 추가, 적용
